chore: update codes with new constant data
