--- app/constants/classroomData.ts
@@ -1,15 +1,0 @@
-export const classroomData = {
-  teacherName: "Mr. John Doe",
-  section: "A",
-  studentCount: 235,
-  books: [
-    { title: "Book 01", author: "Author 01" },
-    { title: "Book 02", author: "Author 02" },
-    { title: "Book 03", author: "Author 03" },
-  ],
-};
-
-export const allowedRoles = {
-  student: "student",
-  teacher: "teacher",
-};
