+++ middleware.ts
@@ -2,13 +2,15 @@ import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
 
 export function middleware(request: NextRequest) {
-  // Check if the user is trying to access the root "/"
-  if (request.nextUrl.pathname === "/") {
-    return NextResponse.redirect(new URL("/login", request.url));
+  const { pathname } = request.nextUrl;
+
+  if (pathname === "/") {
+    return NextResponse.redirect(`${request.nextUrl.origin}/login`);
   }
+
+  return NextResponse.next();
 }
 
-// Config to specify the paths where the middleware should run
 export const config = {
-  matcher: "/", // Apply middleware only to the root path
+  matcher: "/",
 };
